implements Wasm type analysis

+++ sturdy-core/src/main/scala/sturdy/values/floating/FloatOps.scala
@@ -26,12 +26,12 @@ trait FloatOps[B, V]:
   def copysign(v: V, sign: V): V
 
 type ConvertFloatInt[VFrom, VTo] = Convert[Float, Int, VFrom, VTo, config.Overflow && config.Bits]
-type ConvertFloatLong[VFrom, VTo] = Convert[Long, Int, VFrom, VTo, config.Overflow && config.Bits]
+type ConvertFloatLong[VFrom, VTo] = Convert[Float, Long, VFrom, VTo, config.Overflow && config.Bits]
 type ConvertFloatDouble[VFrom, VTo] = Convert[Float, Double, VFrom, VTo, NilCC.type]
 type ConvertFloatBytes[VFrom, VTo] = Convert[Float, Seq[Byte], VFrom, VTo, SomeCC[ByteOrder]]
 type ConvertBytesFloat[VFrom, VTo] = Convert[Seq[Byte], Float, VFrom, VTo, SomeCC[ByteOrder]]
 
-type ConvertDoubleInt[VFrom, VTo] = Convert[Double, Float, VFrom, VTo, config.Overflow && config.Bits]
+type ConvertDoubleInt[VFrom, VTo] = Convert[Double, Int, VFrom, VTo, config.Overflow && config.Bits]
 type ConvertDoubleLong[VFrom, VTo] = Convert[Double, Long, VFrom, VTo, config.Overflow && config.Bits]
 type ConvertDoubleFloat[VFrom, VTo] = Convert[Double, Float, VFrom, VTo, NilCC.type]
 type ConvertDoubleBytes[VFrom, VTo] = Convert[Double, Seq[Byte], VFrom, VTo, SomeCC[ByteOrder]]
